Update Lab03B0405st.java
Changed float to double

+++ Lab 3B 4-5/Lab03B0405st.java
@@ -40,7 +40,7 @@ class Rational
    // Complete implementation required for the 80-Point Version.
 	public double getDecimal() 
    { 
-      return num/ ((float) den);
+      return num/ ((double) den);
    }
    
    // Complete implementation required for the 80-Point Version.
